ISSUE 3: Added some comments in English
