+++ isara/src/main/java/org/iesvi/isara/app/IsaraApplication.java
@@ -3,6 +3,11 @@ package org.iesvi.isara.app;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+/**
+ * Main application, based on Spring Boot, where the service starts.
+ *
+ * @author Isa González
+ */
 @SpringBootApplication
 public class IsaraApplication {
