+++ isara/src/main/java/org/iesvi/isara/shared/config/MyConfiguration.java
@@ -4,6 +4,13 @@ import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * The goal of ModelMapper is to make object mapping easy, by automatically determining how one object model maps
+ * to another, based on conventions, in the same way that a human would - while providing a simple, refactoring-safe API
+ * for handling specific use cases.
+ *
+ * @author Isa González
+ */
 @Configuration
 public class MyConfiguration {
     @Bean
