+++ isara/src/main/java/org/iesvi/isara/userManagement/domain/User.java
@@ -1,13 +1,15 @@
-package org.iesvi.isara.gestionUsuario.domain;
+package org.iesvi.isara.userManagement.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.iesvi.isara.shared.domain.Address;
+import org.iesvi.isara.shared.domain.UserAddress;
 
 /**
  * User parent class with common attributes for all users who will use the application.
+ *
+ * @author Isa González
  */
 @Entity
 @Data
@@ -16,14 +18,14 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id_user", nullable = false)
-    private Integer idUser;
+    private Long idUser;
     private String userName;
     private String password;
     private String firstName;
     private String lastName;
     @ManyToOne
     @JoinColumn(name = "address_id")
-    private Address address;
+    private UserAddress address;
     private String phoneNumber;
     private String email;
 }
