+++ isara/src/main/java/org/iesvi/isara/userManagement/domain/UserAdmin.java
@@ -1,4 +1,4 @@
-package org.iesvi.isara.gestionUsuario.domain;
+package org.iesvi.isara.userManagement.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -7,15 +7,18 @@ import lombok.NoArgsConstructor;
 
 /**
  * Child class that extends the user class.
+ *
+ * @author Isa González
  */
 @Entity
 @Data
 @AllArgsConstructor @NoArgsConstructor
-public class UserCustomer extends User {
+public class UserAdmin extends User {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id_customer", nullable = false)
-    private Integer idCustomer;
+    @Column(name = "id_admin", nullable = false)
+    private Long idAdmin;
+    private Double salary;
 
     @JoinColumn(name = "user_id")
     private Integer userId;
