+++ isara/src/main/java/org/iesvi/isara/userManagement/infra/dto/CreateUserDTO.java
@@ -1,8 +1,12 @@
-package org.iesvi.isara.gestionUsuario.infra.dto;
+package org.iesvi.isara.userManagement.infra.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
+/**
+ *
+ * @author Isa González
+ */
 @Getter
 @Setter
 public class CreateUserDTO {
