+++ isara/src/main/java/org/iesvi/isara/userManagement/infra/dto/UpdateUserAdminDTO.java
@@ -1,17 +1,22 @@
-package org.iesvi.isara.gestionUsuario.infra.dto;
+package org.iesvi.isara.userManagement.infra.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.iesvi.isara.shared.domain.Address;
+import org.iesvi.isara.shared.domain.UserAddress;
 
+/**
+ *
+ * @author Isa González
+ */
 @Getter
 @Setter
-public class UpdateUserDTO {
+public class UpdateUserAdminDTO {
     private String userName;
     private String password;
     private String firstName;
     private String lastName;
-    private Address address;
+    private UserAddress address;
     private String phoneNumber;
     private String email;
+    private Double salary;
 }
