+++ isara/src/main/java/org/iesvi/isara/userManagement/infra/dto/UserDTO.java
@@ -1,18 +1,21 @@
-package org.iesvi.isara.gestionUsuario.infra.dto;
+package org.iesvi.isara.userManagement.infra.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.iesvi.isara.shared.domain.Address;
+import org.iesvi.isara.shared.domain.UserAddress;
 
+/**
+ *
+ * @author Isa González
+ */
 @Getter
 @Setter
-public class UpdateUserAdminDTO {
+public class UserDTO {
     private String userName;
     private String password;
     private String firstName;
     private String lastName;
-    private Address address;
+    private UserAddress address;
     private String phoneNumber;
     private String email;
-    private Double salary;
 }
