+++ isara/src/main/java/org/iesvi/isara/userManagement/infra/dto/converter/UserDTOConverter.java
@@ -1,11 +1,15 @@
-package org.iesvi.isara.gestionUsuario.infra.dto.converter;
+package org.iesvi.isara.userManagement.infra.dto.converter;
 
 import lombok.RequiredArgsConstructor;
-import org.iesvi.isara.gestionUsuario.domain.User;
-import org.iesvi.isara.gestionUsuario.infra.dto.UserDTO;
+import org.iesvi.isara.userManagement.domain.User;
+import org.iesvi.isara.userManagement.infra.dto.UserDTO;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
+/**
+ *
+ * @author Isa González
+ */
 @Component
 @RequiredArgsConstructor
 public class UserDTOConverter {
