+++ isara/src/main/java/org/iesvi/isara/userManagement/infra/persistence/UserCustomerRepository.java
@@ -1,7 +1,10 @@
-package org.iesvi.isara.gestionUsuario.infra.persistence;
+package org.iesvi.isara.userManagement.infra.persistence;
 
-import org.iesvi.isara.gestionUsuario.domain.UserCustomer;
+import org.iesvi.isara.userManagement.domain.UserCustomer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+/**
+ * @author Isa González
+ */
 public interface UserCustomerRepository extends JpaRepository<UserCustomer, Long> {
 }
