+++ isara/src/main/java/org/iesvi/isara/userManagement/infra/persistence/UserRepository.java
@@ -1,0 +1,11 @@
+package org.iesvi.isara.userManagement.infra.persistence;
+
+import org.iesvi.isara.userManagement.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+/**
+ *
+ * @author Isa González
+ */
+public interface UserRepository extends JpaRepository<User, Long> {
+}
